There was a problem with the original CreateProblemResult method. The "contentType": "application/problem+json" and "statusCode": 404 properties viable outside the problemDetails object come from returning the response via IResult (e.g., Results.Problem(...)) from ProcessError method.
ProcessError returns an ActionResult, but the origional method was using Results.Problem(...) (which returns IResult), ASP.NET wraps the IResult into an ObjectResult, adding metadata like statusCode and contentType.

Thus instead of returning IResult from CreateProblemResult, return ProblemDetails, and construct the ObjectResult directly with it inside ProcessError.
In addition, I addee a Time Stamp for when the problem details occur.